New bot runner file, seperated jobs
